moved byte data class
